routerのメソッド指定をhandlerで分割
